Move skaled config methods, update helper-scripts
