+++ tools/logger.py
@@ -32,6 +32,7 @@ from tools.configs.logs import (
     ADMIN_LOG_PATH,
     API_LOG_FORMAT,
     API_LOG_PATH,
+    MIRAGE_LOG_FORMAT,
     SYNC_LOG_PATH,
     DEBUG_LOG_PATH,
     LOG_FILE_SIZE_BYTES,
@@ -112,6 +113,10 @@ def init_logger(log_format, log_file_path=None, debug_file_path=None):
 
 def init_admin_logger():
     init_logger(ADMIN_LOG_FORMAT, ADMIN_LOG_PATH, DEBUG_LOG_PATH)
+
+
+def init_mirage_logger():
+    init_logger(MIRAGE_LOG_FORMAT, ADMIN_LOG_PATH, DEBUG_LOG_PATH)
 
 
 def init_api_logger():
